Simplified sensor state
Simplified sensor state + changes to MQTT logic
