+++ app.js
@@ -4,7 +4,7 @@ var path = require('path');
 var cookieParser = require('cookie-parser');
 var logger = require('morgan');
 var cron = require('./cron');
-var serverMQTT = require('./serverMQTT');
+var serverMQTT = require('./MQTT/serverMQTT');
 
 var indexRouter = require('./routes/index');
 var tempRouter = require('./routes/temps');
@@ -43,12 +43,12 @@ app.use('/mirror', mirror_Router);
 
 
 // catch 404 and forward to error handler
-app.use(function(req, res, next) {
+app.use(function (req, res, next) {
   next(createError(404));
 });
 
 // error handler
-app.use(function(err, req, res, next) {
+app.use(function (err, req, res, next) {
   // set locals, only providing error in development
   res.locals.message = err.message;
   res.locals.error = req.app.get('env') === 'development' ? err : {};
